Overload die to return a hash of the error and the full call stack.
Make the case of a compile error (which appear to not follow $SIG{__DIE__}
catch and create a hash and "original die" that.

Make the error printer inspect the first element of the callstack and
try to pretty print everything.  Added two more error tests:

error0001.asp (existed and checks compile errors)
error0002.asp (new and checks for runtime "in ASP" errors)
error0003.asp (new and checks for runtime errors in a called module)

refs #3